requirements.txt
aenum==2.2.6
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cassandra-driver==3.25.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.7
django-cassandra-engine==1.6.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
geomet==0.2.1.post1
gevent==21.8.0
greenlet==1.1.2
gremlinpython==3.4.6
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isodate==0.6.0
lz4==3.1.3
msgpack==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==21.0.0
python-snappy==0.6.0
pytz==2021.1
requests==2.26.0
scales==1.0.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tornado==5.1.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket==0.2.1
websocket-client==1.2.1
zope.event==4.5.0
zope.interface==5.4.0
